Revoke the movement on State.
